requirements.txt
asgiref==3.5.2
autopep8==1.7.0
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
dash==2.7.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-jwt
djangorestframework-simplejwt
Flask==2.2.2
fonttools==4.38.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
numpy==1.23.4
packaging==21.3
Pillow==9.3.0
plotly==5.11.0
pycodestyle==2.9.1
PyJWT
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
tenacity==8.1.0
toml==0.10.2
Werkzeug==2.2.2
